--- patcherex/arch/cgc/utils.py
@@ -1,1 +1,0 @@
-from patcherex.utils import *
