--- patcherex/arch/i386/utils.py
@@ -1,1 +1,0 @@
-from patcherex.utils import *
