+++ patcherex/backends/detourbackends/arm.py
@@ -1,206 +1,27 @@
-from patcherex.backends.detourbackend import *
-from elftools.elf.elffile import ELFFile
+import bisect
+import logging
+import os
+import re
+from collections import OrderedDict, defaultdict
+
+import capstone
+import keystone
 from elftools.construct.lib import Container
-from . import utils
+from elftools.elf.elffile import ELFFile
 
-class DetourBackend(Backend):
+from patcherex.backend import Backend
+from patcherex.backends.detourbackends._utils import *
+from patcherex.backends.detourbackends._elf import *
+from patcherex.patches import *
+from patcherex.utils import *
+
+l = logging.getLogger("patcherex.backends.DetourBackend")
+
+class DetourBackendArm(DetourBackendElf):
     # how do we want to design this to track relocations in the blocks...
-    def __init__(self, filename, data_fallback=None, base_address=None, try_pdf_removal=True):
-        super(DetourBackend, self).__init__(filename, project_options={"main_opts": {"base_addr": base_address}})
-
-        self.modded_segments = self.dump_segments() # dump_segments also set self.structs
-
-        self.cfg = self._generate_cfg()
-        self.ordered_nodes = self._get_ordered_nodes(self.cfg)
-
-        # header stuff
-        self.ncontent = self.ocontent
-        self.segments = None
-        self.original_header_end = None
-
-        # tag to track if already patched
-        self.patched_tag = b"SHELLPHISH\x00"  # should not be longer than 0x20
-
-        self.name_map = RejectingDict()
-
-        # where to put the segments in memory
+    def __init__(self, filename, base_address=None):
+        super(DetourBackendArm, self).__init__(filename, base_address=base_address)
         self.added_code_segment = 0x0600000
-        self.added_data_segment = 0x07000000
-        current_hdr = self.structs.Elf_Ehdr.parse(self.ncontent)
-        self.single_segment_header_size = current_hdr["e_phentsize"]
-        assert self.single_segment_header_size >= self.structs.Elf_Phdr.sizeof()
-        # we may need up to 3 additional segments (patch code, patch data, phdr)
-        self.additional_headers_size = 3 * self.single_segment_header_size
-
-        self.added_code = b""
-        self.added_data = b""
-        self.added_code_file_start = None
-        self.added_data_file_start = None
-        self.added_patches = []
-        self.added_rwdata_len = 0
-        self.added_rwinitdata_len = 0
-        self.phdr_segment = None
-        self.to_init_data = b""
-        self.max_convertible_address = (1 << (32 if self.structs.elfclass == 32 else 48)) - 1
-
-        self.saved_states = OrderedDict()
-        # not all the touched bytes are bad, they are only a serious problem in case of InsertCodePatch
-        self.touched_bytes = set()
-
-        self.name_map["ADDED_DATA_START"] = (len(self.ncontent) % 0x1000) + self.added_data_segment
-
-    def is_patched(self):
-        return self.ncontent.startswith(self.patched_tag, self.structs.Elf_Ehdr.sizeof())
-
-    def setup_headers(self, segments):
-        #if self.is_patched():
-        #    return
-
-        # copying original program headers (potentially modified by patches)
-        # in the new place (at the  end of the file)
-        self.first_load = None
-        load_segments_rounded = []
-        for segment in segments:
-            if segment["p_type"] == "PT_LOAD":
-                if self.first_load is None:
-                    self.first_load = segment
-                load_segments_rounded.append((
-                        # start of the segment, round down to multiple of 0x1000
-                        (segment["p_vaddr"] - self.first_load["p_vaddr"]) - ((segment["p_vaddr"] - self.first_load["p_vaddr"]) % 0x1000), 
-                        # end of the segment, round up to multiple of 0x1000
-                        int((segment["p_vaddr"] + segment["p_memsz"] - self.first_load["p_vaddr"] + 0x1000 - 1) / 0x1000 * 0x1000) ))
-
-        for segment in segments:
-            if segment["p_type"] == "PT_PHDR":
-                if self.phdr_segment is not None:
-                    raise ValueError("Multiple PHDR segments!")
-                self.phdr_segment = segment
-
-                segment["p_filesz"] += self.additional_headers_size
-                segment["p_memsz"]  += self.additional_headers_size
-
-                phdr_size = max(segment["p_filesz"], segment["p_memsz"])
-
-                load_segments_rounded = sorted(load_segments_rounded, key=lambda x: x[0])
-
-                # combine overlapping load segments
-                while True:
-                    new_load_segments_rounded = []
-                    i = 0
-                    while i < len(load_segments_rounded) - 1:
-                        prev_seg = load_segments_rounded[i]
-                        next_seg = load_segments_rounded[i + 1]
-                        if prev_seg[1] > next_seg[0]: # two segments overlap
-                            new_load_segments_rounded.append((prev_seg[0], next_seg[1])) # append combine of two segments
-                            i += 2
-                        else:
-                            new_load_segments_rounded.append(prev_seg) # append segment without overlap
-                            i += 1
-                    if i == len(load_segments_rounded) - 1:
-                        new_load_segments_rounded.append(load_segments_rounded[i]) # append last segment if without overlapping
-                    if new_load_segments_rounded == load_segments_rounded: # if no overlap
-                        break
-                    load_segments_rounded = new_load_segments_rounded # combined segments, run again
-
-                for prev_seg, next_seg in zip(load_segments_rounded[:-1], load_segments_rounded[1:]):
-                    potential_base = ((max(prev_seg[1], len(self.ncontent)) + 0xF) & ~0xF) # round up to 0x10
-                    if next_seg[0] - potential_base > phdr_size: # if there is space between segments, put phdr here
-                        self.phdr_start = potential_base
-                        break
-                else:
-                    self.phdr_start = load_segments_rounded[-1][1] # otherwise put it after the last load segment
-
-                segment["p_offset"]  = self.phdr_start
-                segment["p_vaddr"]   = self.phdr_start + self.first_load["p_vaddr"]
-                segment["p_paddr"]   = self.phdr_start + self.first_load["p_vaddr"]
-
-        self.ncontent = self.ncontent.ljust(self.phdr_start, b"\x00")
-
-        # change pointer to program headers to point at the end of the elf
-        current_hdr = self.structs.Elf_Ehdr.parse(self.ncontent)
-        old_phoff = current_hdr["e_phoff"]
-        current_hdr["e_phoff"] = len(self.ncontent)
-        self.ncontent = utils.bytes_overwrite(self.ncontent, self.structs.Elf_Ehdr.build(current_hdr), 0)
-
-        print("putting them at %#x" % self.phdr_start)
-        print("current len: %#x" % len(self.ncontent))
-        for segment in segments:
-            if segment["p_type"] == "PT_PHDR":
-                segment = self.phdr_segment
-            self.ncontent = utils.bytes_overwrite(self.ncontent, self.structs.Elf_Phdr.build(segment))
-        self.original_header_end = len(self.ncontent)
-
-        # we overwrite the first original program header,
-        # we do not need it anymore since we have moved original program headers at the bottom of the file
-        self.ncontent = utils.bytes_overwrite(self.ncontent, self.patched_tag, old_phoff)
-
-        # adding space for the additional headers
-        # I add two of them, no matter what, if the data one will be used only in case of the fallback solution
-        # Additionally added program headers have been already copied by the for loop above
-        self.ncontent = self.ncontent.ljust(len(self.ncontent)+self.additional_headers_size, b"\x00")
-
-    def dump_segments(self, tprint=False):
-        with open(self.filename, "rb") as f:
-            elf = ELFFile(f)
-            self.structs = elf.structs
-            segments = []
-            for i in range(elf.num_segments()):
-                seg = elf.get_segment(i)
-                segments.append(seg.header)
-        return segments
-
-    def set_added_segment_headers(self, nsegments):
-        l.debug("added_data_file_start: %#x", self.added_data_file_start)
-        added_segments = 0
-        original_nsegments = nsegments
-
-        # add a LOAD segment for the PHDR segment
-        phdr_segment_header = Container(**{ "p_type":   1,                                      "p_offset": self.phdr_segment["p_offset"],
-                                            "p_vaddr":  self.phdr_segment["p_vaddr"],           "p_paddr":  self.phdr_segment["p_paddr"],
-                                            "p_filesz": self.phdr_segment["p_filesz"],          "p_memsz":  self.phdr_segment["p_memsz"],
-                                            "p_flags":  0x4,                                    "p_align":  0x1000})
-        self.ncontent = utils.bytes_overwrite(self.ncontent, self.structs.Elf_Phdr.build(phdr_segment_header),
-                                                self.original_header_end + (2 * self.structs.Elf_Phdr.sizeof()))
-        added_segments += 1
-
-        # add a LOAD segment for the DATA segment
-        data_segment_header = Container(**{ "p_type":   1,                                      "p_offset": self.added_data_file_start,
-                                            "p_vaddr":  self.name_map["ADDED_DATA_START"],      "p_paddr":  self.name_map["ADDED_DATA_START"],
-                                            "p_filesz": len(self.added_data),                   "p_memsz":  len(self.added_data),
-                                            "p_flags":  0x6,                                    "p_align":  0x1000})
-        self.ncontent = utils.bytes_overwrite(self.ncontent, self.structs.Elf_Phdr.build(data_segment_header),
-                                                self.original_header_end + self.structs.Elf_Phdr.sizeof())
-        added_segments += 1
-
-        # add a LOAD segment for the CODE segment
-        code_segment_header = Container(**{ "p_type":   1,                                      "p_offset": self.added_code_file_start,
-                                            "p_vaddr":  self.name_map["ADDED_CODE_START"],      "p_paddr":  self.name_map["ADDED_CODE_START"],
-                                            "p_filesz": len(self.added_code),                   "p_memsz":  len(self.added_code),
-                                            "p_flags":  0x5,                                    "p_align":  0x1000})
-        self.ncontent = utils.bytes_overwrite(self.ncontent, self.structs.Elf_Phdr.build(code_segment_header),
-                                            self.original_header_end)
-        added_segments += 1
-
-        # print original_nsegments,added_segments
-        current_hdr = self.structs.Elf_Ehdr.parse(self.ncontent)
-        current_hdr["e_phnum"] += added_segments
-        self.ncontent = utils.bytes_overwrite(self.ncontent, self.structs.Elf_Ehdr.build(current_hdr), 0)
-
-    def set_oep(self, new_oep):
-        # set original entry point
-        current_hdr = self.structs.Elf_Ehdr.parse(self.ncontent)
-        current_hdr["e_entry"] = new_oep
-        self.ncontent = utils.bytes_overwrite(self.ncontent, self.structs.Elf_Ehdr.build(current_hdr), 0)
-
-    def get_oep(self):
-        # get original entry point
-        current_hdr = self.structs.Elf_Ehdr.parse(self.ncontent)
-        return current_hdr["e_entry"]
-
-    # 3 inserting strategies
-    # Jump out and jump back
-    # move a single function out
-    # extending all the functions, so all need to move
 
     def get_block_containing_inst(self, inst_addr):
         index = bisect.bisect_right(self.ordered_nodes, inst_addr) - 1
@@ -223,31 +44,6 @@ class DetourBackend(Backend):
                 return self.check_if_thumb(inst_addr + 1)
             else:
                 raise MissingBlockException("Couldn't find a block containing address %#x" % inst_addr)
-
-    def get_current_code_position(self):
-        return self.name_map["ADDED_CODE_START"] + (len(self.ncontent) - self.added_code_file_start)
-
-    def save_state(self,applied_patches):
-        self.saved_states[tuple(applied_patches)] = (self.ncontent,set(self.touched_bytes),self.name_map.copy())
-
-    def restore_state(self,applied_patches,removed_patches):
-        # find longest sequence of patches for which we have a save state
-        if len(removed_patches) > 0:
-            cut = min([len(applied_patches)]+[applied_patches.index(p) for p in removed_patches if p in applied_patches])
-            applied_patches = applied_patches[:cut]
-        current_longest = self.saved_states[tuple(applied_patches)]
-        self.ncontent, self.touched_bytes, self.name_map = current_longest
-        #print "retrieving",applied_patches
-
-        # cut dictionary to the current state
-        todict = OrderedDict()
-        for i, (k, v) in enumerate(self.saved_states.items()):
-            if i > list(self.saved_states.keys()).index(tuple(applied_patches)):
-                break
-            todict[k]=v
-        self.saved_states = todict
-
-        return applied_patches
 
     def apply_patches(self, patches):
         # deal with stackable patches
@@ -311,7 +107,7 @@ class DetourBackend(Backend):
             if isinstance(patch, RemoveInstructionPatch):
                 if patch.ins_size is None:
                     ins = self.read_mem_from_file(patch.ins_addr, 4)
-                    size = utils.disassemble(ins, 0, is_thumb=self.check_if_thumb(patch.ins_addr))[0].size
+                    size = self.disassemble(ins, 0, is_thumb=self.check_if_thumb(patch.ins_addr))[0].size
                 else:
                     size = patch.ins_size
                 self.patch_bin(patch.ins_addr, b"\x00\xbf" * int((size + 2 - 1) / 2) if self.check_if_thumb(patch.ins_addr) else b"\x00\xF0\x20\xE3" * int((size + 4 - 1) / 4))
@@ -346,12 +142,12 @@ class DetourBackend(Backend):
         for patch in patches:
             if isinstance(patch, AddCodePatch):
                 if patch.is_c:
-                    code_len = len(utils.compile_c(patch.asm_code,
+                    code_len = len(self.compile_c(patch.asm_code,
                                                    optimization=patch.optimization,
                                                    compiler_flags=patch.compiler_flags,
                                                    is_thumb=patch.is_thumb))
                 else:
-                    code_len = len(utils.compile_asm(patch.asm_code,
+                    code_len = len(self.compile_asm(patch.asm_code,
                                                                  current_symbol_pos,
                                                                  is_thumb=patch.is_thumb))
                 if patch.name is not None:
@@ -361,12 +157,12 @@ class DetourBackend(Backend):
         for patch in patches:
             if isinstance(patch, AddCodePatch):
                 if patch.is_c:
-                    new_code = utils.compile_c(patch.asm_code,
+                    new_code = self.compile_c(patch.asm_code,
                                                optimization=patch.optimization,
                                                compiler_flags=patch.compiler_flags,
                                                is_thumb=patch.is_thumb)
                 else:
-                    new_code = utils.compile_asm(patch.asm_code,
+                    new_code = self.compile_asm(patch.asm_code,
                                                  self.get_current_code_position(),
                                                  self.name_map,
                                                  is_thumb=patch.is_thumb)
@@ -385,10 +181,10 @@ class DetourBackend(Backend):
                 new_oep = self.get_current_code_position()
                 # only r0 (a1) need to be saved, ref: glibc/sysdeps/{ARCH}/start.S
                 instructions = "push {{r0}}"
-                instructions += patch.asm_code 
+                instructions += patch.asm_code
                 instructions += "pop {{r0}}"
                 instructions += "\nbl {}".format(hex(int(old_oep)))
-                new_code = utils.compile_asm(instructions,
+                new_code = self.compile_asm(instructions,
                                              self.get_current_code_position(),
                                              self.name_map,
                                              is_thumb=patch.is_thumb)
@@ -404,7 +200,7 @@ class DetourBackend(Backend):
             if isinstance(patch, InlinePatch):
                 obj = self.project.loader.main_object
                 prog_origin = patch.instruction_addr if not obj.pic else obj.addr_to_offset(patch.instruction_addr)
-                new_code = utils.compile_asm(patch.new_asm,
+                new_code = self.compile_asm(patch.new_asm,
                                                 prog_origin,
                                                 self.name_map,
                                                 is_thumb=self.check_if_thumb(patch.instruction_addr))
@@ -475,86 +271,22 @@ class DetourBackend(Backend):
                 segments = [segments[0]] + [patch.new_segment] + segments[1:]
 
             self.setup_headers(segments)
-            self.set_added_segment_headers(len(segments))
+            self.set_added_segment_headers()
             l.debug("final symbol table: "+ repr([(k,hex(v)) for k,v in self.name_map.items()]))
         else:
             l.info("no patches, the binary will not be touched")
-
-    def handle_remove_patch(self,patches,patch):
-        # note the patches contains also "future" patches
-        l.info("Handling removal of patch: "+str(patch))
-        cleaned_patches = [p for p in patches if p != patch]
-        removed_patches = [patch]
-        while True:
-            removed = False
-            for p in cleaned_patches:
-                for d in p.dependencies:
-                    if d not in cleaned_patches:
-                        l.info("Removing depending patch: "+str(p)+" depends from "+str(d))
-                        removed = True
-                        if p in cleaned_patches:
-                            cleaned_patches.remove(p)
-                        if p not in removed_patches:
-                            removed_patches.append(p)
-            if removed == False:
-                break
-        return cleaned_patches,removed_patches
 
     def check_if_movable(self, instruction, is_thumb=False):
         # the idea here is an instruction is movable if and only if
         # it has the same string representation when moved at different offsets is "movable"
         def bytes_to_comparable_str(ibytes, offset):
-            return " ".join(utils.instruction_to_str(utils.disassemble(ibytes, offset, is_thumb=is_thumb)[0]).split()[2:])
+            return " ".join(utils.instruction_to_str(self.disassemble(ibytes, offset, is_thumb=is_thumb)[0]).split()[2:])
 
         instruction_bytes = instruction.bytes
         pos1 = bytes_to_comparable_str(instruction_bytes, 0x0)
         pos2 = bytes_to_comparable_str(instruction_bytes, 0x07f00000)
         pos3 = bytes_to_comparable_str(instruction_bytes, 0xfe000000)
         return pos1 == pos2 and pos2 == pos3
-
-    def maddress_to_baddress(self, addr):
-        if addr >= self.max_convertible_address:
-            msg = "%08x higher than max_convertible_address (%08x)" % (addr,self.max_convertible_address)
-            raise InvalidVAddrException(msg)
-        baddr = self.project.loader.main_object.addr_to_offset(addr)
-        if baddr is None:
-            raise InvalidVAddrException(hex(addr))
-        else:
-            return baddr
-
-    def get_memory_translation_list(self, address, size):
-        # returns a list of address ranges that map to a given virtual address and size
-        start = address
-        end = address+size-1  # we will take the byte at end
-        start_p = address & 0xfffffff000
-        end_p = end & 0xfffffff000
-        if start_p == end_p:
-            return [(self.maddress_to_baddress(start), self.maddress_to_baddress(end)+1)]
-        else:
-            first_page_baddress = self.maddress_to_baddress(start)
-            mlist = list()
-            mlist.append((first_page_baddress, (first_page_baddress & 0xfffffff000)+0x1000))
-            nstart = (start & 0xfffffff000) + 0x1000
-            while nstart != end_p:
-                mlist.append((self.maddress_to_baddress(nstart), self.maddress_to_baddress(nstart)+0x1000))
-                nstart += 0x1000
-            mlist.append((self.maddress_to_baddress(nstart), self.maddress_to_baddress(end)+1))
-            return mlist
-
-    def patch_bin(self, address, new_content):
-        # since the content could theoretically be split into different segments we will handle it here
-        ndata_pos = 0
-
-        for start, end in self.get_memory_translation_list(address, len(new_content)):
-            ndata = new_content[ndata_pos:ndata_pos+(end-start)]
-            self.ncontent = utils.bytes_overwrite(self.ncontent, ndata, start)
-            ndata_pos += len(ndata)
-
-    def read_mem_from_file(self, address, size):
-        mem = b""
-        for start, end in self.get_memory_translation_list(address, size):
-            mem += self.ncontent[start : end]
-        return mem
 
     def get_movable_instructions(self, block):
         # TODO there are two improvements here:
@@ -563,7 +295,7 @@ class DetourBackend(Backend):
         # get movable_instructions in the bb
         original_bbcode = block.bytes
         is_thumb = self.check_if_thumb(block.addr)
-        instructions = utils.disassemble(original_bbcode, block.addr, is_thumb=is_thumb)
+        instructions = self.disassemble(original_bbcode, block.addr, is_thumb=is_thumb)
 
         if self.check_if_movable(instructions[-1], is_thumb=is_thumb):
             movable_instructions = instructions
@@ -602,16 +334,16 @@ class DetourBackend(Backend):
     def compile_moved_injected_code(self, classified_instructions, patch_code, offset=0, is_thumb=False):
         # create injected_code (pre, injected, culprit, post, jmp_back)
         injected_code = "_patcherex_begin_patch:\n"
-        injected_code += "\n".join([utils.capstone_to_asm(i)
+        injected_code += "\n".join([self.capstone_to_asm(i)
                                     for i in classified_instructions
                                     if i.overwritten == 'pre'])
         injected_code += "\n"
         injected_code += patch_code + "\n"
-        injected_code += "\n".join([utils.capstone_to_asm(i)
+        injected_code += "\n".join([self.capstone_to_asm(i)
                                     for i in classified_instructions
                                     if i.overwritten == 'culprit'])
         injected_code += "\n"
-        injected_code += "\n".join([utils.capstone_to_asm(i)
+        injected_code += "\n".join([self.capstone_to_asm(i)
                                     for i in classified_instructions
                                     if i.overwritten == 'post'])
         injected_code += "\n"
@@ -626,7 +358,7 @@ class DetourBackend(Backend):
         injected_code = "\n".join([line for line in injected_code.split("\n") if line != ""])
         l.debug("injected code:\n%s", injected_code)
 
-        compiled_code = utils.compile_asm(injected_code,
+        compiled_code = self.compile_asm(injected_code,
                                               base=self.get_current_code_position(),
                                               name_map=self.name_map, is_thumb=is_thumb)
         return compiled_code
@@ -681,10 +413,10 @@ class DetourBackend(Backend):
 
         # insert the jump detour
         offset = self.project.loader.main_object.mapped_base if self.project.loader.main_object.pic else 0
-        detour_jmp_code = utils.compile_jmp(detour_pos, self.get_current_code_position() + offset, is_thumb=self.check_if_thumb(detour_pos))
+        detour_jmp_code = self.compile_jmp(detour_pos, self.get_current_code_position() + offset, is_thumb=self.check_if_thumb(detour_pos))
         self.patch_bin(detour_pos, detour_jmp_code)
         patched_bbcode = self.read_mem_from_file(block_addr, block.size)
-        patched_bbinstructions = utils.disassemble(patched_bbcode, block_addr, is_thumb=self.check_if_thumb(block_addr))
+        patched_bbinstructions = self.disassemble(patched_bbcode, block_addr, is_thumb=self.check_if_thumb(block_addr))
         l.debug("patched bb instructions:\n %s",
                 "\n".join([utils.instruction_to_str(i) for i in patched_bbinstructions]))
 
@@ -692,19 +424,93 @@ class DetourBackend(Backend):
 
         return new_code
 
-    def get_final_content(self):
-        return self.ncontent
+    def capstone_to_asm(self, instruction):
+            return instruction.mnemonic + " " + instruction.op_str.replace('{','{{').replace('}','}}')
 
-    def save(self, filename=None):
-        if filename is None:
-            filename = self.filename + "_patched"
+    def disassemble(self, code, offset=0x0, is_thumb=False):
+        md = capstone.Cs(capstone.CS_ARCH_ARM, capstone.CS_MODE_THUMB if is_thumb else capstone.CS_MODE_ARM)
+        md.detail = True
+        if isinstance(code, str):
+            code = bytes(map(ord, code))
+        return list(md.disasm(code, offset))
 
-        final_content = self.get_final_content()
-        with open(filename, "wb") as f:
-            f.write(final_content)
+    def compile_jmp(self, origin, target, is_thumb=False):
+        jmp_str = '''
+            bl {target}
+        '''.format(**{'target': hex(int(target))})
+        return self.compile_asm(jmp_str, base=origin, is_thumb=is_thumb)
 
-        os.chmod(filename, 0o755)
+    def compile_asm(self, code, base=None, name_map=None, is_thumb=False):
+        #print "=" * 10
+        #print code
+        #if base != None: print hex(base)
+        #if name_map != None: print {k: hex(v) for k,v in name_map.iteritems()}
+        try:
+            if name_map is not None:
+                code = code.format(**name_map)  # compile_asm
+            else:
+                code = re.subn(r'{.*?}', "0x41414141", code)[0]  # solve symbols
+        except KeyError as e:
+            raise UndefinedSymbolException(str(e))
+        try:
+            ks = keystone.Ks(keystone.KS_ARCH_ARM, keystone.KS_MODE_THUMB if is_thumb else keystone.KS_MODE_ARM)
+            encoding, count = ks.asm(code, base)
+        except keystone.KsError as e:
+            print("ERROR: %s" %e) #TODO raise some error
+        return bytes(encoding)
 
+    @staticmethod
+    def get_c_function_wrapper_code(function_symbol, get_return=False, debug=False):
+        # TODO maybe with better calling convention on llvm this can be semplified
+        wcode = []
+        # wcode.append("pusha")
+        # TODO add param list handling, right two params in ecx/edx are supported
+        '''
+        assert len(param_list) <= 2 # TODO support more parameters
+        if len(param_list) == 1:
+            wcode.append("mov ecx, %s" % param_list[0])
+        if len(param_list) == 2:
+            wcode.append("mov ecx, %s" % param_list[0])
+            wcode.append("mov edx, %s" % param_list[1])
+        '''
+        # if debug:
+        #     wcode.append("int 0x3")
+        wcode.append("bl {%s}" % function_symbol)
+        # if get_return:
+        #     wcode.append("mov [esp+28], eax") #FIXME check
+        # wcode.append("popa")
 
-def init_backend(program_name, options):
-    return DetourBackend(program_name, **options)
+        return "\n".join(wcode)
+
+    def compile_c(self, code, optimization='-Oz', name_map=None, compiler_flags="-m32", is_thumb=False):
+        # TODO symbol support in c code
+        with utils.tempdir() as td:
+            c_fname = os.path.join(td, "code.c")
+            object_fname = os.path.join(td, "code.o")
+            bin_fname = os.path.join(td, "code.bin")
+
+            fp = open(c_fname, 'w')
+            fp.write(code)
+            fp.close()
+
+            res = utils.exec_cmd("clang -nostdlib -mno-sse -target arm-linux-gnueabihf -ffreestanding %s %s -o %s -c %s %s" \
+                            % ("-mthumb" if is_thumb else "", optimization, object_fname, c_fname, compiler_flags), shell=True)
+            if res[2] != 0:
+                print("CLang error:")
+                print(res[0])
+                print(res[1])
+                fp = open(c_fname, 'r')
+                fcontent = fp.read()
+                fp.close()
+                print("\n".join(["%02d\t%s"%(i+1,l) for i,l in enumerate(fcontent.split("\n"))]))
+                raise CLangException
+            res = utils.exec_cmd("arm-linux-gnueabihf-objcopy -O binary -j .text %s %s" % (object_fname, bin_fname), shell=True)
+            if res[2] != 0:
+                print("objcopy error:")
+                print(res[0])
+                print(res[1])
+                raise ObjcopyException
+            fp = open(bin_fname, "rb")
+            compiled = fp.read()
+            fp.close()
+        return compiled
